Move slice to ast.property_accessors
