+++ pyccel/codegen/printing/ccode.py
@@ -33,7 +33,9 @@ from pyccel.ast.literals  import LiteralString, LiteralInteger, Literal, Nil
 from pyccel.ast.numpyext import NumpyFull, NumpyArray
 from pyccel.ast.numpyext import NumpyReal, NumpyImag, NumpyFloat
 
-from pyccel.ast.variable import Slice, ValuedVariable
+from pyccel.ast.property_accessors import Slice
+
+from pyccel.ast.variable import ValuedVariable
 from pyccel.ast.variable import PyccelArraySize, Variable, VariableAddress
 from pyccel.ast.variable import DottedName
