+++ pyccel/parser/semantic.py
@@ -43,7 +43,7 @@ from pyccel.ast.core import EmptyNode
 from pyccel.ast.variable import Constant
 from pyccel.ast.variable import Variable
 from pyccel.ast.variable import TupleVariable
-from pyccel.ast.variable import Slice, IndexedElement
+from pyccel.ast.variable import IndexedElement
 from pyccel.ast.variable import DottedName, DottedVariable
 from pyccel.ast.variable import ValuedVariable
 from pyccel.ast.core import ValuedArgument
@@ -90,6 +90,8 @@ from pyccel.ast.numpyext import NumpyInt, NumpyInt32, NumpyInt64
 from pyccel.ast.numpyext import NumpyFloat, NumpyFloat32, NumpyFloat64
 from pyccel.ast.numpyext import NumpyComplex, NumpyComplex64, NumpyComplex128
 from pyccel.ast.numpyext import NumpyArrayClass, NumpyNewArray
+
+from pyccel.ast.property_accessors import Slice
 
 from pyccel.ast.sympy_helper import sympy_to_pyccel, pyccel_to_sympy
