+++ pyccel/parser/syntactic.py
@@ -64,7 +64,9 @@ from pyccel.ast.headers  import Header, MetaVariable
 from pyccel.ast.literals import LiteralInteger, LiteralFloat, LiteralComplex
 from pyccel.ast.literals import LiteralFalse, LiteralTrue, LiteralString, Nil
 from pyccel.ast.functionalexpr import FunctionalSum, FunctionalMax, FunctionalMin
-from pyccel.ast.variable  import DottedName, Slice
+from pyccel.ast.variable  import DottedName
+
+from pyccel.ast.property_accessors import Slice
 
 from pyccel.parser.extend_tree import extend_tree
 from pyccel.parser.base import BasicParser
